Delete interfaces where they do not bring a clear benefit (or are even confusing)
 - restructure code in packages
