+++ types/namespaceid.go
@@ -1,0 +1,13 @@
+package types
+
+import "bytes"
+
+type NamespaceID []byte
+
+func (nid NamespaceID) Less(other NamespaceID) bool {
+	return bytes.Compare(nid, other) < 0
+}
+
+func (nid NamespaceID) Equal(other NamespaceID) bool {
+	return bytes.Equal(nid, other)
+}
